project3/eatgo/eatgo-admin-api/src/test/java/com/myepark/eatgo/application/RegionServiceTest.java
package com.myepark.eatgo.application;

import com.myepark.eatgo.domain.Region;
import com.myepark.eatgo.domain.RegionRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.ArrayList;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.BDDMockito.given;

public class RegionServiceTest {

    @InjectMocks
    private RegionService regionService;

    @Mock
    private RegionRepository regionRepository;

    @BeforeEach
    void beforeEach() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void getRegions() {
        List<Region> resource = new ArrayList<>();
        resource.add(Region.builder().name("Seoul").build());

        given(regionRepository.findAll())
                .willReturn(resource);

        List<Region> regions = regionService.getRegions();

        Region region = regions.get(0);
        assertThat(region.getName()).isEqualTo("Seoul");
    }

}